pyproject.toml
[project]
name = "legal-rag-api"
version = "0.1.0"
description = "Legal Document Q&A — RAG system with FastAPI + Celery + pgvector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.3",
    "greenlet>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
    "minio>=7.2.20",
    "python-multipart>=0.0.26",
    "pymupdf>=1.27.2.2",
    "openai>=2.31.0",
    "celery[rabbitmq]>=5.6.3",
    "langchain-text-splitters>=1.1.1",
    "langchain-openai>=1.1.13",
    "pgvector>=0.3.0",
    "redis>=7.4.0",
    "cohere>=5.21.1",
    "langchain-cohere>=0.5.1",
]
